# service/docs/swagger.yaml
definitions:
...
  generate.GenerateHtmlRequest:
    properties:
      template:
        example: https://drive.google.com/uc?export=download&id=16oauTQqVnJtJEl8unMYyUpH6BILRS97C
        type: string
      variables:
        additionalProperties:
          type: string
        example:
          amount: "100.00"
          currency: EUR
          date: "2025-05-04"
          name: John Doe
          reference: ABC123
        type: object
    required:
    - template
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/generate/html:
    post:
      consumes:
      - application/json
      description: Generate a PDF document from template based on the provided variables
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/generate.GenerateHtmlRequest'
      produces:
      - application/pdf
      responses:
        "200":
          description: PDF file
          schema:
            type: file
        "400":
          description: Invalid input
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Generate something
      tags:
      - PDF Generate
...
swagger: "2.0"